keychain: Move dcrwallet impl to dcrwallet pkg
This commit moves the dcrwallet-backed implementation of the KeyRing and
SecretKeyRing keychain interfaces from that package to the
lnwallet/dcrwallet package.

This uncouples the keychain package from dcrwallet, which allows future
implementations to use modules that themselves use hdkeychain without
causing import cycles. It also improves abstraction and possible package
reuse by not having it depend on internal wallet packages and
concentrates all dcrwallet-related functionality into a single package.

The implementation itself of the WalletKeyRing is the same. The
call sites were changed to use the new location for it.

In order to allow moving the implementation to a different package, the
interface tests were refactored to export check functions that are
callable from client packages. This allows the keychain interface tests
to be ran from external packages so that they can ensure they conform to
the required semantics.
